requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
google-genai>=0.3.0
Pillow>=10.0.0
flask>=3.0.0
gunicorn>=21.0.0
